Made UI improvements to the photos tab and profile footer
- moved the profile footer to the bottom of the body tag in a footer
tag to improve code structure and style
- added padding to the bottom of the main body content of each tab
to accommodate for the padding of the footer. This fixed the issue where
the footer covered a lot of the bottom data on the photos,
destinations and trips tabs
- redesigned the photos tab page and removed the small scrolling card
container in favour of just displaying the photos on the main photos tab.
This allows the main page scroll to be utilized instead of the hard to use
mini scroll from the card.

#story[571]